refactor: sun 모듈로 변경
